+++ R/runPDTMTptmAnalysis.R
@@ -24,8 +24,16 @@
 #' @param proteinIdColProteins,proteinIdColPeptides Character strings pointing
 #'     to columns of the \code{rowData} of the respective objects that
 #'     contain the protein identifiers (will be used to match the two objects).
+#' @param modificationsCol Character string pointing to a column of the
+#'     \code{rowData} of the peptides object that contain modification
+#'     details. \code{excludeUnmodifiedPeptides} and \code{keepModifications}
+#'     will use information from this column.
 #' @param excludeUnmodifiedPeptides Logical scalar, whether to filter out
 #'     peptides without modifications.
+#' @param keepModifications Character string (or \code{NULL}) indicating
+#'     which modifications to retain in the analysis. Can be a regular
+#'     expression, which will be matched against the \code{modificationsCol}.
+#'     If \code{NULL} (the default), all rows are retained.
 #' @param comparisons List of character vectors defining comparisons to
 #'     perform. The first element of each vector represents the
 #'     denominator of the comparison. If not empty, \code{ctrlGroup} and
@@ -98,7 +106,8 @@ runPDTMTptmAnalysis <- function(
         assayForTests, assayImputation,
         proteinIdColProteins = function(df) einprot::getFirstId(df, "einprotProtein", ";"),
         proteinIdColPeptides = function(df) einprot::getFirstId(df, "einprotProtein", ";"),
-        excludeUnmodifiedPeptides = TRUE, comparisons = list(),
+        modificationsCol = "Modifications", excludeUnmodifiedPeptides = TRUE,
+        keepModifications = NULL, comparisons = list(),
         ctrlGroup = "", allPairwiseComparisons = TRUE, singleFit = FALSE,
         subtractBaseline = FALSE, baselineGroup = "",
         testType = "interaction", minNbrValidValues = 2,
@@ -121,7 +130,9 @@ runPDTMTptmAnalysis <- function(
         assayForTests = assayForTests, assayImputation = assayImputation,
         proteinIdColProteins = proteinIdColProteins,
         proteinIdColPeptides = proteinIdColPeptides,
+        modificationsCol = modificationsCol,
         excludeUnmodifiedPeptides = excludeUnmodifiedPeptides,
+        keepModifications = keepModifications,
         comparisons = comparisons, ctrlGroup = ctrlGroup,
         allPairwiseComparisons = allPairwiseComparisons, singleFit = singleFit,
         subtractBaseline = subtractBaseline, baselineGroup = baselineGroup,
@@ -151,7 +162,9 @@ runPDTMTptmAnalysis <- function(
              assayForTests = assayForTests, assayImputation = assayImputation,
              proteinIdColProteins = proteinIdColProteins,
              proteinIdColPeptides = proteinIdColPeptides,
+             modificationsCol = modificationsCol,
              excludeUnmodifiedPeptides = excludeUnmodifiedPeptides,
+             keepModifications = keepModifications,
              comparisons = comparisons, ctrlGroup = ctrlGroup,
              allPairwiseComparisons = allPairwiseComparisons, singleFit = singleFit,
              subtractBaseline = subtractBaseline, baselineGroup = baselineGroup,